Better support newer projects install docs
The ops:docs-install-guide tag as currently written does not allow a project
which has made a guide using source, rather than packages, to attain the tag.

This a particular problem for new projects and creates a catch-22 situation
where distros are only likely to package more 'mature'/'adopted' projects,
but users might only adopt a project if there's an install guide.

This change removes the references to package management systems while keeping
the need for installation guides to support installation (whether it be source,
packages, or some other method) on the multiple distros required to make a
project installable to the majority of OpenStack users.

Change-Id: I939ea9643e9d609f2ab6592fd68e0b735f53f252
Co-Authored-By: Mike Perez <thingee@gmail.com>